Mosaic/requirements.txt
xarray
pandas
geopandas
rioxarray
rasterio
h5netcdf
flask
pytest
fastparquet
pyarrow
zarr
uvicorn[standard]
gunicorn
dask
odc-geo
scipy